requirements.txt
crowdstrike-falconpy>=1.2.11
tls-syslog>=0.2.0
pymongo>=4.2.0
